(process_mips_specific): Print conflictsno as an unsigned long.
From: binutils CVS HEAD branch, based on rev 1.231
Provided by: Daniel Lucq <daniel at lucq dot org>

ok fgsch@ drahn@

+++ gnu/usr.bin/binutils/binutils/readelf.c
@@ -8144,7 +8144,8 @@ process_mips_specific (file)
 	    iconf[cnt] = BYTE_GET (econf64[cnt]);
 	}
 
-      printf (_("\nSection '.conflict' contains %d entries:\n"), conflictsno);
+      printf (_("\nSection '.conflict' contains %lu entries:\n"),
+	      (unsigned long) conflictsno);
       puts (_("  Num:    Index       Value  Name"));
 
       for (cnt = 0; cnt < conflictsno; ++cnt)
